refactor windows
